Rust compiles!

+++ rust/src/ixgbe/driver_constants.rs
@@ -4,6 +4,6 @@ pub const RING_SIZE: usize = 256;
 
 pub const MAX_OUTPUTS: usize = 256;
 
-pub const FLUSH_PERIOD: u8 = 8;
+pub const FLUSH_PERIOD: usize = 8;
 
 pub const RECYCLE_PERIOD: u8 = 64;
